feat: Replace data source with resource
