+++ .changeset/silent-hornets-sit.md
@@ -2,4 +2,4 @@
 "provider-azure": minor
 ---
 
-Added new data source and new resource to retrieve the first available subnet CIDR within a specified Azure Virtual Network
+Added new resource to retrieve the first available subnet CIDR within a specified Azure Virtual Network
