--- infra/provider/docs/data-sources/available_subnet_cidr.md
@@ -1,39 +1,0 @@
----
-# generated by https://github.com/hashicorp/terraform-plugin-docs
-page_title: "dx_available_subnet_cidr Data Source - terraform-provider-azure"
-subcategory: ""
-description: |-
-  Finds an available CIDR block for a new subnet within a specified Azure Virtual Network.
----
-
-# dx_available_subnet_cidr (Data Source)
-
-Find an available CIDR block for a new subnet within a specified Azure Virtual Network. This data source automatically calculates a non-overlapping CIDR within the address space of the VNet, avoiding existing subnets.
-
-## Example Usage
-
-```terraform
-data "dx_available_subnet_cidr" "next_cidr" {
-  virtual_network_id = azurerm_virtual_network.example.id
-  prefix_length      = 24
-}
-
-resource "azurerm_subnet" "new_subnet" {
-  name                 = "example-subnet"
-  resource_group_name  = azurerm_resource_group.example.name
-  virtual_network_name = azurerm_virtual_network.example.name
-  address_prefixes     = [data.dx_available_subnet_cidr.next_cidr.cidr_block]
-}
-```
-
-<!-- schema generated by tfplugindocs -->
-## Schema
-
-### Required
-
-- `prefix_length` (Number) The desired prefix length for the new subnet CIDR (e.g., 24 for /24). Must be larger than the VNet prefix and smaller or equal to 29.
-- `virtual_network_id` (String) The Azure Resource ID of the Virtual Network.
-
-### Read-Only
-
-- `cidr_block` (String) The calculated available CIDR block.
